requirements.txt
python-telegram-bot[job-queue,webhooks]==20.7
requests